Move transfer information from Information.ts to Payment.ts
The function to get transfer information for identity wallets was
not tested before. This is fixed now.
